fix: resolve compilation errors in repo and handler
- Use correct import aliases (dbuser, dbent) in user_repo.go
- Fix response helper function names (ParsePagination, Paginated)
- Use strconv.ParseInt for ID params instead of non-existent helpers
- Fix int/int64 type mismatch in pagination result
- Rename toDTO to toAdminInviteCodeDTO to avoid conflicts

+++ backend/internal/repository/admin_invite_code_repo.go
@@ -96,7 +96,7 @@ func (r *adminInviteCodeRepo) List(ctx context.Context, params pagination.Pagina
 		result[i] = *toServiceAdminInviteCode(c)
 	}
 	return result, &pagination.PaginationResult{
-		Total:    total,
+		Total:    int64(total),
 		Page:     params.Page,
 		PageSize: params.PageSize,
 		Pages:    (total + params.PageSize - 1) / params.PageSize,
